seine/bootstrap: generate and use an ansible-seine package
playbooks may wish to install ansible and it should then be kept
in the image. To support this use-case, an ansible-seine package
is generated using equivs. It is then installed and ansible pulled
as a dependency. The generated package is removed using apt-get
autoremove ansible-seine to let the package manager determine if
ansible should be kept in the generated system.

Signed-off-by: Cedric Hombourger <chombourger@gmail.com>